VCPU: Rewrite register caching and access
- Add a RegisterProtocol to ensure the Registers types have the correct
  members in both versions.

- Add a readRegisters() call to read the registers from the VMCS which
  can fail if the underlying vmcs becomes invalid.

- Keep tracks of which registers are updated so that setting the
  registers before vmentry only sets the necessary registers.

- Fix the readNatural()/writeNatural() to use UInt64, the natural size
  relates to the data in the value not the vmread/vmwrite datatype.

+++ Sources/VMMKit/vmexit.swift
@@ -151,7 +151,7 @@ public enum VMExit: Equatable {
         public let writeable: Bool
         public let executable: Bool
         public let guestPhysicalAddress: PhysicalAddress
-        public let guestLinearAddress: UInt?
+        public let guestLinearAddress: UInt64?
 
         public var description: String {
             let perms = (readable ? "r" : "-") + (writeable ? "w" : "-") + (executable ? "x" : "-")
